+++ CHANGELOG.md
@@ -37,14 +37,18 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 - Removed gorilla mux library and replaced it with net/http updates in go 1.22.
 - Clean up `ProposeBlock` for validator client to reduce cognitive scoring and enable further changes.
 - Updated k8s-io/client-go to v0.30.4 and k8s-io/apimachinery to v0.30.4
+- Migrated tracing library from opencensus to opentelemetry for both the beacon node and validator.
+- Refactored light client code to make it more readable and make future PRs easier.
 
 ### Deprecated
 - `--disable-grpc-gateway` flag is deprecated due to grpc gateway removal.
 - `--enable-experimental-state` flag is deprecated. This feature is now on by default. Opt-out with `--disable-experimental-state`.
 
 ### Removed
+
 - removed gRPC Gateway
 - Removed unused blobs bundle cache
+- Removed consolidation signing domain from params. The Electra design changed such that EL handles consolidation signature verification.
 
 ### Fixed
 
@@ -58,6 +62,7 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 - `== nil` checks before calling `IsNil()` on interfaces to prevent panics.
 - Core: Fixed slash processing causing extra hashing
 - Core: Fixed extra allocations when processing slashings
+- remove unneeded container in blob sidecar ssz response
 
 ### Security